+++ Middlewares/ST/STM32_USB_Device_Library/Core/Src/usbd_core.c
@@ -513,9 +513,11 @@ USBD_StatusTypeDef USBD_ClrClassConfig(USBD_HandleTypeDef *pdev, uint8_t cfgidx)
   }
 #else
   /* Clear configuration  and De-initialize the Class process */
-  if (pdev->pClass[0]->DeInit(pdev, cfgidx) != 0U)
-  {
-    ret = USBD_FAIL;
+  if(pdev->pClass[0] != NULL) {
+    if (pdev->pClass[0]->DeInit(pdev, cfgidx) != 0U)
+    {
+      ret = USBD_FAIL;
+    }
   }
 #endif /* USE_USBD_COMPOSITE */
 
@@ -862,7 +864,11 @@ USBD_StatusTypeDef USBD_LL_SetSpeed(USBD_HandleTypeDef *pdev,
 
 USBD_StatusTypeDef USBD_LL_Suspend(USBD_HandleTypeDef *pdev)
 {
-  pdev->dev_old_state = pdev->dev_state;
+  if (pdev->dev_state != USBD_STATE_SUSPENDED)
+  {
+    pdev->dev_old_state = pdev->dev_state;
+  }
+
   pdev->dev_state = USBD_STATE_SUSPENDED;
 
   return USBD_OK;
@@ -1123,20 +1129,21 @@ uint8_t USBD_CoreFindEP(USBD_HandleTypeDef *pdev, uint8_t index)
   * @param  pdev: device instance
   * @param  ep_dir: USBD_EP_IN or USBD_EP_OUT
   * @param  ep_type: USBD_EP_TYPE_CTRL, USBD_EP_TYPE_ISOC, USBD_EP_TYPE_BULK or USBD_EP_TYPE_INTR
+  * @param  ClassId: The Class ID
   * @retval Address of the selected endpoint or 0xFFU if no endpoint found.
   */
-uint8_t USBD_CoreGetEPAdd(USBD_HandleTypeDef *pdev, uint8_t ep_dir, uint8_t ep_type)
+uint8_t USBD_CoreGetEPAdd(USBD_HandleTypeDef *pdev, uint8_t ep_dir, uint8_t ep_type, uint8_t ClassId)
 {
   uint8_t idx;
 
   /* Find the EP address in the selected class table */
-  for (idx = 0; idx < pdev->tclasslist[pdev->classId].NumEps; idx++)
+  for (idx = 0; idx < pdev->tclasslist[ClassId].NumEps; idx++)
   {
-    if (((pdev->tclasslist[pdev->classId].Eps[idx].add & USBD_EP_IN) == ep_dir) && \
-        (pdev->tclasslist[pdev->classId].Eps[idx].type == ep_type) && \
-        (pdev->tclasslist[pdev->classId].Eps[idx].is_used != 0U))
+    if (((pdev->tclasslist[ClassId].Eps[idx].add & USBD_EP_IN) == ep_dir) && \
+        (pdev->tclasslist[ClassId].Eps[idx].type == ep_type) && \
+        (pdev->tclasslist[ClassId].Eps[idx].is_used != 0U))
     {
-      return (pdev->tclasslist[pdev->classId].Eps[idx].add);
+      return (pdev->tclasslist[ClassId].Eps[idx].add);
     }
   }