+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 v0.2.3
-- 
+- Add USB composite interface (CDC + MIDI)